Move follow sources to reader/follow-sources
