+++ client/reader/follow-sources.jsx
