+++ client/reader/following-manage/feed-search-results.jsx
@@ -15,7 +15,7 @@ import classnames from 'classnames';
  */
 import Button from 'components/button';
 import ReaderSubscriptionListItemPlaceholder from 'blocks/reader-subscription-list-item/placeholder';
-import { READER_FOLLOWING_MANAGE_SEARCH_RESULT } from 'reader/follow-button/follow-sources';
+import { READER_FOLLOWING_MANAGE_SEARCH_RESULT } from 'reader/follow-sources';
 import InfiniteStream from 'components/reader-infinite-stream';
 import { siteRowRenderer } from 'components/reader-infinite-stream/row-renderers';
 import { requestFeedSearch } from 'state/reader/feed-searches/actions';
