+++ client/reader/following-manage/subscriptions.jsx
@@ -28,7 +28,7 @@ import EllipsisMenu from 'components/ellipsis-menu';
 import PopoverMenuItem from 'components/popover/menu-item';
 import { formatUrlForDisplay, getFeedTitle } from 'reader/lib/feed-display-helper';
 import { addQueryArgs } from 'lib/url';
-import { READER_SUBSCRIPTIONS } from 'reader/follow-button/follow-sources';
+import { READER_SUBSCRIPTIONS } from 'reader/follow-sources';
 
 class FollowingManageSubscriptions extends Component {
 	static propTypes = {
