+++ client/reader/stream/empty-search-recommended-post.jsx
@@ -9,7 +9,7 @@ import React from 'react';
  */
 import { RelatedPostCard } from 'blocks/reader-related-card-v2';
 import { recordTrackForPost, recordAction } from 'reader/stats';
-import { EMPTY_SEARCH_RECOMMENDATIONS } from 'reader/follow-button/follow-sources';
+import { EMPTY_SEARCH_RECOMMENDATIONS } from 'reader/follow-sources';
 
 export default function EmptySearchRecommendedPost( { post } ) {
 	function handlePostClick() {
